db/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "db"
version = "0.1.0"
description = "Общая БД для бота и WebApp (модели, сессия, репозитории)"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0",
    "aiosqlite>=0.19",
    "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["db*"]
